+++ src/libponyrt/sched/systematic_testing.h
@@ -5,9 +5,6 @@
 #include "scheduler.h"
 
 #if defined(USE_SYSTEMATIC_TESTING)
-#if !defined(PLATFORM_IS_WINDOWS) && !defined(USE_SCHEDULER_SCALING_PTHREADS)
-pony_static_assert(false, "Systematic testing requires pthreads (USE_SCHEDULER_SCALING_PTHREADS) to be enabled!");
-#endif
 
 #include <stdio.h>
 
@@ -24,33 +21,23 @@ size_t ponyint_systematic_testing_static_alloc_size();
 #endif
 
 void ponyint_systematic_testing_init(uint64_t random_seed, uint32_t max_threads);
-void ponyint_systematic_testing_start(scheduler_t* schedulers, pony_thread_id_t pinned_actor_thread, pony_signal_event_t pinned_actor_signal);
-void ponyint_systematic_testing_wait_start(pony_thread_id_t thread, pony_signal_event_t signal);
+void ponyint_systematic_testing_start();
+void ponyint_systematic_testing_wait_start(int32_t scheduler_index);
 void ponyint_systematic_testing_yield();
 // Current value of the systematic-testing logical clock (see the .c file). Used
 // by the scheduler in place of ponyint_cpu_tick() on the clock-gated control
 // paths so those decisions are a deterministic function of the seed.
 uint64_t ponyint_systematic_testing_clock();
 void ponyint_systematic_testing_stop_thread();
-#if defined(USE_SCHEDULER_SCALING_PTHREADS)
-void ponyint_systematic_testing_suspend(pthread_mutex_t* mut);
-#else
-void ponyint_systematic_testing_suspend();
-#endif
 
 // TODO systematic testing: maybe make output conditional based on a `verbosity` argument that is CLI driven?
 #define SYSTEMATIC_TESTING_PRINTF(f_, ...) fprintf(stderr, (f_), ##__VA_ARGS__)
 #define SYSTEMATIC_TESTING_INIT(RANDOM_SEED, MAX_THREADS) ponyint_systematic_testing_init(RANDOM_SEED, MAX_THREADS)
-#define SYSTEMATIC_TESTING_START(SCHEDULERS, PINNED_ACTOR_THREAD, PINNED_ACTOR_SLEEP_OBJECT) ponyint_systematic_testing_start(SCHEDULERS, PINNED_ACTOR_THREAD, PINNED_ACTOR_SLEEP_OBJECT)
-#define SYSTEMATIC_TESTING_WAIT_START(THREAD, SIGNAL) ponyint_systematic_testing_wait_start(THREAD, SIGNAL)
+#define SYSTEMATIC_TESTING_START() ponyint_systematic_testing_start()
+#define SYSTEMATIC_TESTING_WAIT_START(SCHEDULER_INDEX) ponyint_systematic_testing_wait_start(SCHEDULER_INDEX)
 #define SYSTEMATIC_TESTING_YIELD() ponyint_systematic_testing_yield()
 #define SYSTEMATIC_TESTING_CLOCK() ponyint_systematic_testing_clock()
 #define SYSTEMATIC_TESTING_STOP_THREAD() ponyint_systematic_testing_stop_thread()
-#if defined(USE_SCHEDULER_SCALING_PTHREADS)
-#define SYSTEMATIC_TESTING_SUSPEND(MUT) ponyint_systematic_testing_suspend(MUT)
-#else
-#define SYSTEMATIC_TESTING_SUSPEND() ponyint_systematic_testing_suspend()
-#endif
 
 PONY_EXTERN_C_END
 
@@ -58,16 +45,11 @@ PONY_EXTERN_C_END
 
 #define SYSTEMATIC_TESTING_PRINTF(f_, ...)
 #define SYSTEMATIC_TESTING_INIT(RANDOM_SEED, MAX_THREADS)
-#define SYSTEMATIC_TESTING_START(SCHEDULERS, PINNED_ACTOR_THREAD, PINNED_ACTOR_SLEEP_OBJECT)
-#define SYSTEMATIC_TESTING_WAIT_START(THREAD, SIGNAL)
+#define SYSTEMATIC_TESTING_START()
+#define SYSTEMATIC_TESTING_WAIT_START(SCHEDULER_INDEX)
 #define SYSTEMATIC_TESTING_YIELD()
 #define SYSTEMATIC_TESTING_CLOCK() (0)
 #define SYSTEMATIC_TESTING_STOP_THREAD()
-#if defined(USE_SCHEDULER_SCALING_PTHREADS)
-#define SYSTEMATIC_TESTING_SUSPEND(MUT)
-#else
-#define SYSTEMATIC_TESTING_SUSPEND()
-#endif
 
 #endif